Update README.md
The question is in the form "How" rather than "Can I..." so direct answer seems to be confusing. I consider at least two ways of handling this confusion - remove the direct answer (which I did in this PR) or change the question to take shape of the answer (i.e. 'Can I access Redux store outside a component').